requirements.txt
requests
python-dotenv
urllib3
pytest
tabulate
ollama
